# src/config/swagger.json
{
  "openapi": "3.0.3",
  "servers": [{ "url": "" }],
  "info": {
    "title": "Monix Documentation API",
    "version": "1.0",
    "description": "La documentation du backend de monix (le logiciel de gestion de crédits du club nix)"
  },
...
  "paths": {
...
    "/users/": {
      "get": {
        "summary": "Get all users",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "Reception de la liste d'utilisateurs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserList"
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create a user",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserCreation"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "Création d'un utilisateur",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGet"
                }
              }
            }
          }
        }
      }
    },
    "/users/me": {
      "get": {
        "summary": "Get current user",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "L'utilisateur a bien été recupéré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGet"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update current user",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserModificationMe"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "L'utilisateur a bien été modifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGet"
                }
              }
            }
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "summary": "Get one user",
        "tags": ["Users"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true,
            "description": "ID de l'utilisateur dont on veut les détails"
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "L'utilisateur a bien été recupéré",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGet"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update one user",
        "tags": ["Users"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true,
            "description": "ID de l'utilisateur que l'on veut modifier"
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UserModification"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "L'utilisateur a bien été modifié",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGet"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete one user",
        "tags": ["Users"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": { "type": "integer" },
            "required": true,
            "description": "ID de l'utilisateur que l'on veut modifier"
          }
        ],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "L'utilisateur a bien été supprimé",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiMessage"
                }
              }
            }
          }
        }
      }
    },
    "/users/uploadAvatar": {
      "post": {
        "summary": "Upload an avatar for the currently connected user",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "avatar": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "L'avatar de l'utilisateur à bien été mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserGet"
                }
              }
            }
          }
        }
      }
    },
    "/users/changePassword": {
      "post": {
        "summary": "Let the connected user change his own password",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordBody"
              }
            }
          }
        },
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "L'avatar de l'utilisateur à bien été mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiMessage"
                }
              }
            }
          }
        }
      }
    },
    "/users/generateCode": {
      "post": {
        "summary": "Create a personnal code for the user (connection with code)",
        "tags": ["Users"],
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "default": {
            "description": "Une erreur a eu lieu",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiError"
                }
              }
            }
          },
          "200": {
            "description": "Le code de l'utilisateur à bien été mis à jour",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/codeCreationResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "apiError": {
        "type": "object",
        "required": ["success", "error"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "error": {
            "type": "string"
          }
        }
      },
      "apiMessage": {
        "type": "object",
        "required": ["success", "message"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        }
      },
...
      "codeCreationResponse": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": {
            "type": "boolean"
          },
          "data": {
            "type": "object",
            "required": ["code"],
            "properties": {
              "code": { "type": "number" }
            }
          }
        }
      },
      "User": {
        "type": "object",
        "required": [
          "id",
          "username",
          "email",
          "balance",
          "avatar",
          "createdAt",
          "updatedAt"
        ],
        "properties": {
          "id": { "type": "number" },
          "username": { "type": "string" },
          "email": { "type": "string" },
          "balance": { "type": "number" },
          "avatar": { "type": "string" },
          "createdAt": { "type": "string" },
          "updatedAt": { "type": "string" }
        }
      },
      "UserCreation": {
        "type": "object",
        "required": ["username", "email", "password", "balance"],
        "properties": {
          "username": { "type": "string" },
          "email": { "type": "string" },
          "password": { "type": "string" },
          "balance": { "type": "number" },
          "admin": { "type": "boolean" }
        }
      },
      "UserModificationMe": {
        "type": "object",
        "required": ["username", "email"],
        "properties": {
          "username": { "type": "string" },
          "email": { "type": "string" }
        }
      },
      "UserModification": {
        "type": "object",
        "required": ["username", "email", "balance", "admin"],
        "properties": {
          "username": { "type": "string" },
          "email": { "type": "string" },
          "balance": { "type": "number" },
          "admin": { "type": "string" }
        }
      },
      "UserGet": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": { "type": "boolean" },
          "data": { "$ref": "#/components/schemas/User" }
        }
      },
      "UserList": {
        "type": "object",
        "required": ["success", "data"],
        "properties": {
          "success": { "type": "boolean" },
          "data": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/User" }
          }
        }
      },
      "ChangePasswordBody": {
        "type": "object",
        "required": ["oldPassword", "newPassword", "passwordConfirmation"],
        "properties": {
          "oldPassword": { "type": "string" },
          "newPassword": { "type": "string" },
          "passwordConfirmation": { "type": "string" }
        }
      },
...
    }
  }
}